Number of Minimums on a Segment
Statement : 
Now change the code of the segment tree so that, in addition to the minimum on a segment, it also counts the number of elements equal to the minimum.

Input
The first line contains two integers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The next line contains 𝑛 numbers 𝑎𝑖, the initial state of the array (0≤𝑎𝑖≤109). The following lines contain the description of the operations. The description of each operation is as follows:

1 𝑖 𝑣: set the element with index 𝑖 to 𝑣 (0≤𝑖<𝑛, 0≤𝑣≤109).
2 𝑙 𝑟: calculate the minimum and number of elements equal to minimum of elements with indices from 𝑙 to 𝑟−1 (0≤𝑙<𝑟≤𝑛).
Output
For each operation of the second type print two integers: the minimum on a segment, and the number of elements equal to minimum.

Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>