Segment with the Maximum Sum
Statement : 
In this problem, you need to write a segment tree to find the segment with the maximum sum.

Input
The first line contains two numbers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The next line contains 𝑛 numbers 𝑎𝑖, the initial state of the array (−109≤𝑎𝑖≤109). The following lines contain the description of the operations. The description of each operation is as follows: 𝑖 𝑣, assign the value 𝑣 to the element with index 𝑖 (0≤𝑖<𝑛, −109≤𝑣≤109).

Output
Print 𝑚+1 lines: the maximum sum of numbers on a segment before all operations and after each operation. Please note that this segment may be empty (so the sum on it will be equal to 0).



Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>