K-th one
Problem Statement : 
In this problem, you need to add to the segment tree the operation of finding the 𝑘-th one.

Input
The first line contains two numbers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The next line contains 𝑛 numbers 𝑎𝑖, the initial state of the array (𝑎𝑖∈{0,1} ). The following lines contain the description of the operations. The description of each operation is as follows:

1 𝑖: change the element with index 𝑖 to the opposite.
2 𝑘: find the 𝑘-th one (ones are numbered from 0, it is guaranteed that there are enough ones in the array).
Output
For each operation of the second type, print the index of the corresponding one (all indices in this problem are from 0).

Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>