First element at least X - 2
Problem Statement : 
In this task, you need to add to the segment tree the operation of finding for the given 𝑥 and 𝑙 the minimum index 𝑗 such that 𝑗≥𝑙 and 𝑎[𝑗]≥𝑥.

Input
The first line contains two integers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The next line contains 𝑛 numbers 𝑎𝑖, the initial state of the array (0≤𝑎𝑖≤109). The following lines contain the description of the operations. The description of each operation is as follows:

1 𝑖 𝑣: change the item with index 𝑖 to 𝑣 (0≤𝑖<𝑛, 0≤𝑣≤109).
2 𝑥 𝑙: find the minimum index 𝑗 such that 𝑗≥𝑙 and 𝑎[𝑗]≥𝑥 (0≤𝑥≤109, 0≤𝑙<𝑛). If there is no such element, print −1. Indices start from 0.
Output
For each operation of the second type, print the answer for the query.



Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>