First element at least X
Statement : 
In this task, you need to add to the segment tree the operation of finding the minimum index 𝑗 such that 𝑎[𝑗]≥𝑥.

Input
The first line contains two integers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The next line contains 𝑛 numbers 𝑎𝑖, the initial state of the array (0≤𝑎𝑖≤109). The following lines contain the description of the operations. The description of each operation is as follows:

1 𝑖 𝑣: change the item with index 𝑖 to 𝑣 (0≤𝑖<𝑛, 0≤𝑣≤109).
2 𝑥: find the minimum index 𝑗 such that 𝑎[𝑗]≥𝑥. If there is no such element, print −1. Indices start from 0.
Output
For each operation of the second type, print the answer for the query.

Example
inputCopy
5 7
1 3 2 4 6
2 2
2 5
1 2 5
2 4
2 8
1 3 7
2 6
outputCopy
1
4
2
-1
3


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>