Nested Segments
Statement : 
Given an array of 2𝑛 numbers, each number from 1 to 𝑛 in it occurs exactly twice. We say that the segment 𝑦 is nested inside the segment 𝑥 if both occurrences of the number 𝑦 are between the occurrences of the number 𝑥. Find for each segment 𝑖 how many segments that are nested inside it.

Input
The first line contains the number 𝑛 (1≤𝑛≤105), the second line contains 2𝑛 numbers. It is guaranteed that every number from 1 to 𝑛 occurs exactly twice.

Output
Print 𝑛 numbers, the 𝑖-th number is equal to the number of segments nested inside the segment 𝑖.

Example
inputCopy
5
5 1 2 2 3 1 3 4 5 4
outputCopy
1 0 0 0 3 


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>