Intersecting Segments
Statement : 
Given an array of 2𝑛 numbers, each number from 1 to 𝑛 in it occurs exactly twice. We say that the segment 𝑦 intersects the segment 𝑥 if exactly one occurrence of the number 𝑦 is between the occurrences of the number 𝑥. Find for each segment 𝑖 how many segments there are that intersect with it.

Input
The first line contains the number 𝑛 (1≤𝑛≤105), the second line contains 2𝑛 numbers. It is guaranteed that every number from 1 to 𝑛 occurs exactly twice.

Output
Print 𝑛 numbers, the 𝑖-th number is equal to the number of segments that intersect with the segment 𝑖.

Example
input
5
5 1 2 2 3 1 3 4 5 4
output
1 0 1 1 1 


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>