Addition to Segment
Statement : 
There is an array of 𝑛 elements, initially filled with zeros. You need to write a data structure that processes two types of queries:

add to the segment from 𝑙 to 𝑟−1 the number 𝑣,
find the current value of element 𝑖.
You may have heard that such requests can be made using the tree of segments with mass change operations (we will talk about it in the next lesson), but this problem can be solved using a regular segment tree.

Input
The first line contains two numbers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The following lines contain the description of the operations. The description of each operation is as follows:

1 𝑙 𝑟 𝑣: add the value 𝑣 to the segment from 𝑙 to 𝑟−1 (0≤𝑙<𝑟≤𝑛, 0≤𝑣≤109).
2 𝑖: find the value of the element with index 𝑖 (0≤𝑖<𝑛).
Output
For each operation of the second type, print the corresponding value.

Example
input
5 5
1 0 3 3
2 1
1 2 4 4
2 3
2 4
output
3
4
0


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>