Sign alternation
Implement data structure of 𝑛 elements 𝑎1,𝑎2…𝑎𝑛, with the following operations:

assign the element 𝑎𝑖 value 𝑗;
find alternating sign sum in the range from 𝑙 to 𝑟 inclusive (𝑎𝑙−𝑎𝑙+1+𝑎𝑙+2−…±𝑎𝑟).
Input
The first line of the input file contains a natural number 𝑛 (1≤𝑛≤105) — the length of the array. The second line contains the initial values of the elements (non-negative integers not exceeding 104).

The third line contains a positive integer 𝑚 (1≤𝑚≤105) — the number of operations. The following 𝑚 lines contain operations:

the operation of the first type is given by three numbers 0 i j (1≤𝑖≤𝑛, 1≤𝑗≤104).
an operation of the second type is given by three numbers 1 l r (1≤𝑙≤𝑟≤𝑛).
Output
For each operation of the second type, print on a separate line the corresponding sign alternating sum.

Example
inputCopy
3
1 2 3
5
1 1 2
1 1 3
1 2 3
0 2 1
1 1 3
outputCopy
-1
2
-1
3


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>