Cryptography
Given 𝑛 matrices 𝐴1,𝐴2,…,𝐴𝑛 of size 2×2. You need to calculate the product of the matrices 𝐴𝑖,𝐴𝑖+1,…,𝐴𝑗 for several queries. All calculations are performed modulo 𝑟.

Input
The first line of the input file contains integers 𝑟 (1≤𝑟≤10000), 𝑛 (1≤𝑛≤200000), and 𝑚 (1≤𝑚≤200000). The following 𝑛 blocks are two lines each containing two integers, the matrix descriptions. Then follow 𝑚 pairs of integers from 1 to 𝑛, requests for a product on a segment.

Output
Print 𝑚 blocks of two lines, two integers in each: products on segments. Separate blocks with an empty string. All calculations are performed modulo 𝑟.

Example
inputCopy
3 4 4
0 1
0 0

2 1
1 2

0 0
0 2

1 0
0 2

1 4
2 3
1 3
2 2
outputCopy
0 2
0 0

0 2
0 1

0 1
0 0

2 1
1 2



Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>