Number of Different on Segment
Given an array 𝑎, consisting of small integers (1≤𝑎𝑖≤40). You need to build a data structure that processes two types of queries:

find the number of different elements on a segment,
change the element of the array.
Input
The first line contains two integers 𝑛 and 𝑞, the length of the array and the number of queries, respectively (1≤𝑛,𝑞≤105).

The second line contains 𝑛 numbers 𝑎1, ..., 𝑎𝑛, where 𝑎𝑖 is the initial state of the array (1≤𝑎𝑖≤40).

The following 𝑞 lines describe the queries. Each of these lines has the format "𝑡𝑦𝑝𝑒𝑖 𝑥𝑖 𝑦𝑖".

If 𝑡𝑦𝑝𝑒𝑖=1, then in the 𝑖-th query you need to find the number of different elements on a segment from 𝑥𝑖 to 𝑦𝑖, inclusive (in this case 1≤𝑥𝑖≤𝑦𝑖≤𝑛).

If 𝑡𝑦𝑝𝑒𝑖=2, then the element with the index 𝑥𝑖 is set to 𝑦𝑖 (in this case 1≤𝑥𝑖≤𝑛, 1≤𝑦𝑖≤40).

Output
For each request of type 1 print the answer to this request on a separate line.

Example
inputCopy
7 6
1 2 3 6 5 4 19
1 1 3
1 2 5
1 2 4
2 2 8
1 1 6
1 1 3
outputCopy
3
4
3
6
3


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>