Earthquakes
Statement : 
A city is a sequence of 𝑛 cells numbered from 0 to 𝑛−1. Initially, all cells are empty. Then 𝑚 events of one of two types occur sequentially:

a building with the strength ℎ is being constructed in the cell 𝑖 (if the building already existed in this cell, it is demolished and replaced with a new one),
in the interval from 𝑙 to 𝑟−1 an earthquake of power 𝑝 happens, it destroys all buildings whose strength is not more than 𝑝.
Your task is for each earthquake to say how many buildings it will destroy.

Input
The first line contains the numbers 𝑛 and 𝑚, the number of cells and the number of events (1≤𝑛,𝑚≤105). The following 𝑚 lines describe events. The description of each event is as follows:

1 𝑖 ℎ: a building with the strength ℎ is constructed in the cell 𝑖 (0≤𝑖<𝑛, 1≤ℎ≤109).
2 𝑙 𝑟 𝑝: an earthquake of power 𝑝 happens on a segment from 𝑙 to 𝑟−1 (0≤𝑙<𝑟≤𝑛, 0≤𝑝≤109).
Output
For each event of the second type, print how many buildings were destroyed.

Example
inputCopy
5 9
1 0 3
1 2 5
2 0 4 3
1 1 4
1 2 7
2 1 3 6
1 3 8
1 4 4
2 0 5 10
outputCopy
1
1
3


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>