Addition to Segment
This problem is exactly the same as in the previous lesson, but this time try to solve it using the technique that was discussed in the lecture.

There is an array of 𝑛 elements, initially filled with zeros. You need to write a data structure that processes two types of queries:

add to the segment from 𝑙 to 𝑟−1 the number 𝑣,
find the current value of element 𝑖.
Input
The first line contains two numbers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The following lines contain the description of the operations. The description of each operation is as follows:

1 𝑙 𝑟 𝑣: add the value 𝑣 to the segment from 𝑙 to 𝑟−1 (0≤𝑙<𝑟≤𝑛, 0≤𝑣≤109).
2 𝑖: find the value of the element with index 𝑖 (0≤𝑖<𝑛).
Output
For each operation of the second type, print the corresponding value.

Example
inputCopy
5 5
1 0 3 3
2 1
1 2 4 4
2 3
2 4
outputCopy
3
4
0


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>