A. Assignment and Maximal Segment
There is an array of 𝑛 elements, initially filled with zeros. You need to write a data structure that processes two types of queries:

assign value 𝑣 to all elements on the segment from 𝑙 to 𝑟−1,
find the segment with maximal sum.
Input
The first line contains two numbers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The following lines contain the description of the operations. The description of each operation is as follows: 𝑙 𝑟 𝑣: assign the value 𝑣 to all elements on the interval from 𝑙 to 𝑟−1 (0≤𝑙<𝑟≤𝑛, −109≤𝑣≤109).

Output
Print 𝑚 lines: the maximum sum of numbers on a segment after each operation. Please note that this segment may be empty (so the sum on it will be equal to 0).

Example
inputCopy
5 3
0 5 3
1 3 -1
3 4 -5
outputCopy
15
7
3



Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>