B. Inverse and K-th one
There is an array of 𝑛 booleans, initially filled with zeros. You need to write a data structure that processes two types of queries:

change the value of all elements on a segment from 𝑙 to 𝑟−1 to the opposite,
find the index of the 𝑘-th one.
Input
The first line contains two numbers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The following lines contain the description of the operations. The description of each operation is as follows:

1 𝑙 𝑟: change the value of all elements on a segment from 𝑙 to 𝑟−1 to the opposite (0≤𝑙<𝑟≤𝑛),
2 𝑘: the index of the 𝑘-th one (ones are numbered from 0, it is guaranteed that there are enough ones in the array).
Output
For each operation of the second type, print the index of the corresponding one (all indices in this problem are from 0).

Example
inputCopy
5 7
1 1 3
2 1
1 0 2
2 0
2 1
1 0 5
2 2
outputCopy
2
0
2
4



Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>