Create C. Addition and First element at least X.cpp
There is an array of 𝑛 elements, initially filled with zeros. You need to write a data structure that processes two types of queries:

add 𝑣 to all elements on the segment from 𝑙 to 𝑟−1,
find the minimum index 𝑗 such that 𝑗≥𝑙 and 𝑎[𝑗]≥𝑥.
Input
The first line contains two numbers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The following lines contain the description of the operations. The description of each operation is as follows:

1 𝑙 𝑟 𝑣: add 𝑣 to all elements on the segment from 𝑙 to 𝑟−1 (0≤𝑖<𝑛, 0≤𝑣≤104).
2 𝑥 𝑙: find the minimum index 𝑗 such that 𝑗≥𝑙 and 𝑎[𝑗]≥𝑥 (0≤𝑥≤109, 0≤𝑙<𝑛). If there is no such element, print −1.
Indices start from 0.
Output
For each operation of the second type, print the answer for the query.

Example
inputCopy
5 7
1 2 5 3
1 1 3 4
2 3 0
2 5 0
1 4 5 5
2 5 3
2 8 1
outputCopy
1
2
4
4


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>