A. Assignment, Addition, and Sum
There is an array of 𝑛 elements, initially filled with zeros. You need to write a data structure that processes three types of queries:

assign value 𝑣 to all elements on the segment from 𝑙 to 𝑟−1,
add 𝑣 to all elements on the segment from 𝑙 to 𝑟−1,
find the sum on the segment from 𝑙 to 𝑟−1.
Input
The first line contains two numbers 𝑛 and 𝑚 (1≤𝑛,𝑚≤100000), the size of the array and the number of operations. The following lines contain the description of the operations. The description of each operation is as follows:

1 𝑙 𝑟 𝑣: assign 𝑣 to the segment from 𝑙 to 𝑟−1 (0≤𝑙<𝑟≤𝑛, 0≤𝑣≤105).
2 𝑙 𝑟 𝑣: add 𝑣 to the segment from 𝑙 to 𝑟−1 (0≤𝑙<𝑟≤𝑛, 0≤𝑣≤105).
3 𝑙 𝑟: find the sum on the segment from 𝑙 to 𝑟−1 (0≤𝑙<𝑟≤𝑛).
Output
For each operation of the third type, print the corresponding value.

Example
input
5 7
1 0 3 3
2 2 4 2
3 1 3
2 1 5 1
1 0 2 2
3 0 3
3 3 5
output
10
4


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>