B. Add Arithmetic Progression On Segment
You are given an array 𝑥, consisting of 𝑛 elements equal to 0, and 𝑚 queries of two types:

add an arithmetic progression to a segment: the query is described with four integers 𝑙,𝑟,𝑎,𝑑 — for each 𝑙≤𝑖≤𝑟 you should perform 𝑥𝑖+=𝑎+𝑑⋅(𝑖−𝑙);
print current value of a given element.
Input
The first line contains two integers 𝑛,𝑚 (1≤𝑛,𝑚≤2⋅105).

The next 𝑚 lines describe queries. Each line starts with integer 𝑡 (1≤𝑡≤2) — type of the operation. For operations of the first type, four integers follow: 𝑙,𝑟,𝑎,𝑑 (1≤𝑙≤𝑟≤𝑛,1≤𝑎,𝑑≤2⋅105). For operations of the second time one integer between 1 and 𝑛 follows.

Output
For each query of the second type, print the answer.

Example
inputCopy
9 10
1 4 8 1 2
2 6
2 8
1 1 9 3 1
2 6
2 8
1 2 5 2 3
2 1
2 2
2 3
outputCopy
5
9
13
19
3
6
10


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>