C - K-th Sum
There are two arrays 𝑎 and 𝑏, each of which consists of 𝑛 integers. For each pair of numbers (𝑖,𝑗):1≤𝑖,𝑗≤𝑛, write out the sum 𝑎𝑖+𝑏𝑗. Find in the resulting set of sums the 𝑘-th in ascending order.

Input
The first line contains integers 𝑛 and 𝑘 (1≤𝑛≤105, 1≤𝑘≤𝑛2). The second line contains the elements of array 𝑎, the third line contains the elements of array 𝑏. All array elements are positive integers, no more than 109.

Output
Print one number, the required 𝑘-th sum.

Example
inputCopy
5 10
4 2 6 4 8
7 3 1 9 5
outputCopy
9


Signed-off-by: once-ilovesabnaj <161951610+once-ilovesabnaj@users.noreply.github.com>